backend/pyproject.toml
[project]
name = "botverse"
version = "0.1.0"
description = ""
authors = [ 
    {name = "Kashan Haider",email = "hkashan.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (==0.115.9)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "prisma (>=0.15.0,<0.16.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "pinecone (>=6.0.2,<7.0.0)",
    "pinecone[grpc] (>=6.0.2,<7.0.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "chromadb (>=1.0.7,<2.0.0)",
    "langchain-google-genai (>=2.1.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "botverse", from = "src"}]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

